adicionando-desafio1-e-parte-do-2

+++ semana7/aula25/labe-nnusers/src/components/ListaUsuarios.js
@@ -1,6 +1,8 @@
 import React from "react";
 import axios from "axios";
 import styled from "styled-components";
+import Cadastro from "./Cadastro";
+import DetalheUsuarios from "./DetalheUsuarios";
 
 const DeletarBotao= styled.span`
     color: red;
@@ -17,13 +19,19 @@ const TituloLista = styled.p`
    color: #0C0A0D;
    text-align: center;
 `
-
+const BotaoInfo = styled.button`
+    color: #0C0A0D;
+    display: inline;
+    margin: 10px;
+`
 
 
 export default class ListaUsuarios extends React.Component{
-
+ 
     state = {
-        listaUsuarios: []
+        listaUsuarios: [],
+        maisInfo: false,
+        usuarioSelecionado: {}
     }
 
     componentDidMount = () => {
@@ -47,23 +55,45 @@ export default class ListaUsuarios extends React.Component{
         })  
     };
 
-    deletarUsuario = (userId)=> {
-        axios.delete(
-            `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${userId}`,
-            {
-                headers: {
-                    Authorization: "anapatricia-monteiro-dumont",
-                },
+    verPerfil = () =>{
+        this.setState({maisInfo: !this.state.maisInfo})
+    }
+
+    pegarPerfil = (userId) => {
+        this.verPerfil()
+        axios.get(`https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${userId}`,
+        {
+            headers:{
+                Authorization: "anapatricia-monteiro-dumont"
             }
-        )
-        .then((response)=>{
-            alert("Usuário deletado!")
-            this.mostrarUsuario();
-        })
-        .catch((error)=>{
+        }).then(response => {
+            this.setState({usuarioSelecionado: response.data})
+        }).catch(error => {
             console.log(error.message)
         })
     }
+
+    deletarUsuario = (userId)=> {
+        const confirm = window.confirm("Tem certeza que deseja deletar?")
+        if(confirm === true){
+            axios.delete(
+                `https://us-central1-labenu-apis.cloudfunctions.net/labenusers/users/${userId}`,
+                {
+                    headers: {
+                        Authorization: "anapatricia-monteiro-dumont",
+                    },
+                }
+            )
+            .then((response)=>{
+                window.confirm("Tem certeza que deseja deletar?")
+                this.mostrarUsuario();
+            }).catch((error)=>{
+                console.log(error.message)
+            })
+        }
+        
+    }
+
 
     render(){
         const atualizaLista = this.state.listaUsuarios.map((user) => {
@@ -75,6 +105,7 @@ export default class ListaUsuarios extends React.Component{
                 <DeletarBotao onClick ={() => this.deletarUsuario(user.id)}>
                     X
                 </DeletarBotao>
+
             </div>
             );
         });
@@ -82,7 +113,8 @@ export default class ListaUsuarios extends React.Component{
         return(
             <ContainerUsuario>
                 <TituloLista>Lista de Usuários Cadastrados</TituloLista>
-                <TituloLista>{atualizaLista}</TituloLista>
+                <TituloLista> {atualizaLista} </TituloLista>
+                
             </ContainerUsuario>
         )
     }
